Add operator support to e2e tests
This adds a --use_operator flag to the e2e tests. The e2e tests
expect a different component to store a manifest for deployment
in the $ISTIO/install/kubernetes directory.

One aspect of this test case is incorrect. Deletion of the operator
appears not to work within the test framework. The namespace locks
up with the namespace in the terminating state.

+++ tests/e2e/framework/kubernetes.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"net/url"
 	"os"
@@ -55,6 +56,7 @@ const (
 	mcAuthInstallFileNamespace     = "istio-auth-multicluster.yaml"
 	mcRemoteInstallFile            = "istio-remote.yaml"
 	mcSplitHorizonInstallFile      = "istio-multicluster-split-horizon.yaml"
+	authOperatorInstallFile        = "istio-operator.yaml"
 	istioSystem                    = "istio-system"
 	istioIngressServiceName        = "istio-ingress"
 	istioIngressLabel              = "ingress"
@@ -64,7 +66,7 @@ const (
 	defaultSidecarInjectorFile     = "istio-sidecar-injector.yaml"
 	ingressCertsName               = "istio-ingress-certs"
 	maxDeploymentRolloutTime       = 960 * time.Second
-	maxValidationReadyCheckTime    = 30 * time.Second
+	maxValidationReadyCheckTime    = 120 * time.Second
 	maxCNIDeployTime               = 10 * time.Second
 	helmServiceAccountFile         = "helm-service-account.yaml"
 	istioHelmInstallDir            = istioInstallDir + "/helm/istio"
@@ -128,6 +130,7 @@ var (
 	useGalleyConfigValidator = flag.Bool("use_galley_config_validator", true, "Use galley configuration validation webhook")
 	installer                = flag.String("installer", "kubectl", "Istio installer, default to kubectl, or helm")
 	useMCP                   = flag.Bool("use_mcp", true, "use MCP for configuring Istio components")
+	useOperator              = flag.Bool("use_operator", false, "use Operator to deploy Istio components")
 	outboundTrafficPolicy    = flag.String("outbound_trafficpolicy", "ALLOW_ANY", "Istio outbound traffic policy, default to ALLOW_ANY")
 	enableEgressGateway      = flag.Bool("enable_egressgateway", false, "enable egress gateway, default to false")
 	useCNI                   = flag.Bool("use_cni", false,
@@ -212,6 +215,9 @@ func getClusterWideInstallFile() string {
 		if *useMCP {
 			if *authSdsEnable {
 				istioYaml = authSdsInstallFile
+			} else
+			if *useOperator {
+				istioYaml = authOperatorInstallFile
 			} else {
 				istioYaml = authInstallFile
 			}
@@ -551,8 +557,8 @@ func (k *KubeInfo) Teardown() error {
 			}
 		}
 
+		var istioYaml string
 		if *clusterWide {
-			var istioYaml string
 			if *multiClusterDir != "" {
 				if *authEnable {
 					istioYaml = mcAuthInstallFileNamespace
@@ -562,13 +568,30 @@ func (k *KubeInfo) Teardown() error {
 			} else {
 				istioYaml = getClusterWideInstallFile()
 			}
-
+		}
+		if *useOperator {
+			// Need an operator unique delete procedure
+			if _, err := util.Shell("kubectl -n %s get IstioControlPlane example-istiocontrolplane -o=json | jq '.metadata.finalizers = null' | kubectl -n %s apply --kubeconfig=%s -f -",
+				 "istio-operator", "istio-operator", k.KubeConfig); err != nil {
+				log.Errorf("Failed to turn off operator finalizer.")
+				return err
+			}
+			if _, err := util.Shell("kubectl delete ns %s --grace-period=0 --force --kubeconfig=%s",
+				"istio-operator", k.KubeConfig); err != nil {
+				log.Errorf("Failed to delete istio-operator namespace.")
+				return err
+			}
+			if _, err := util.Shell("kubectl delete ns %s --grace-period=0 --force --kubeconfig=%s",
+				k.Namespace, k.KubeConfig); err != nil {
+				log.Errorf("Failed to delete istio-system namespace.")
+				return err
+			}
+		} else {
 			testIstioYaml := filepath.Join(k.TmpDir, "yaml", istioYaml)
-
 			if err := util.KubeDelete(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
 				log.Infof("Safe to ignore resource not found errors in kubectl delete -f %s", testIstioYaml)
 			}
-		} else {
+
 			if err := util.DeleteNamespace(k.Namespace, k.KubeConfig); err != nil {
 				log.Errorf("Failed to delete namespace %s", k.Namespace)
 				return err
@@ -605,7 +628,16 @@ func (k *KubeInfo) Teardown() error {
 	validatingWebhookConfigurationExists := false
 	log.Infof("Deleting namespace %v", k.Namespace)
 	for attempts := 1; attempts <= maxAttempts; attempts++ {
-		namespaceDeleted, _ = util.NamespaceDeleted(k.Namespace, k.KubeConfig)
+		if *useOperator {
+			if _, err := util.Shell("kubectl delete ns istio-operator --grace-period=0 --force",
+				k.Namespace, k.KubeConfig); err != nil {
+				log.Errorf("Failed to delete istio-operator namespace.")
+				return err
+			}
+			namespaceDeleted = true
+		} else {
+			namespaceDeleted, _ = util.NamespaceDeleted(k.Namespace, k.KubeConfig)
+		}
 		// As validatingWebhookConfiguration "istio-galley" will
 		// be delete by kubernetes GC controller asynchronously,
 		// we need to ensure it's deleted before return.
@@ -725,6 +757,31 @@ func (k *KubeInfo) deepCopy(src map[string][]string) map[string][]string {
 	return newMap
 }
 
+func (k *KubeInfo) fileCopy(src string, dst string) (error) {
+        sourceFileStat, err := os.Stat(src)
+        if err != nil {
+                return err
+        }
+
+        if !sourceFileStat.Mode().IsRegular() {
+                return fmt.Errorf("%s is not a regular file", src)
+        }
+
+        source, err := os.Open(src)
+        if err != nil {
+                return err
+        }
+        defer source.Close()
+
+        destination, err := os.Create(dst)
+        if err != nil {
+                return err
+        }
+        defer destination.Close()
+        _, err = io.Copy(destination, source)
+        return err
+}
+
 func (k *KubeInfo) deployIstio() error {
 	istioYaml := nonAuthInstallFileNamespace
 	if *multiClusterDir != "" {
@@ -752,12 +809,6 @@ func (k *KubeInfo) deployIstio() error {
 		if *useGalleyConfigValidator {
 			return errors.New("cannot enable useGalleyConfigValidator in one namespace tests")
 		}
-	}
-
-	// Create istio-system namespace
-	if err := util.CreateNamespace(k.Namespace, k.KubeConfig); err != nil {
-		log.Errorf("Unable to create namespace %s: %s", k.Namespace, err.Error())
-		return err
 	}
 
 	// Deploy the CNI if enabled
@@ -790,41 +841,62 @@ func (k *KubeInfo) deployIstio() error {
 
 	}
 
-	// Apply istio-init
-	yamlDir := filepath.Join(istioInstallDir, initInstallFile)
-	baseIstioYaml := filepath.Join(k.ReleaseDir, yamlDir)
-	testIstioYaml := filepath.Join(k.TmpDir, "yaml", istioYaml)
-	if err := k.generateIstio(baseIstioYaml, testIstioYaml); err != nil {
-		log.Errorf("Generating istio-init.yaml")
-		return err
-	}
-
-	if err := util.KubeApply(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
-		log.Errorf("istio-init.yaml  %s deployment failed", testIstioYaml)
-		return err
-	}
-
-	// TODO(sdake): need a better synchronization
-	time.Sleep(20 * time.Second)
-
-	// Apply main manifest
-	yamlDir = filepath.Join(istioInstallDir, istioYaml)
-	baseIstioYaml = filepath.Join(k.ReleaseDir, yamlDir)
-	testIstioYaml = filepath.Join(k.TmpDir, "yaml", istioYaml)
-
-	if err := k.generateIstio(baseIstioYaml, testIstioYaml); err != nil {
-		log.Errorf("Generating yaml %s failed", testIstioYaml)
-		return err
-	}
-
-	if *multiClusterDir != "" {
-		if err := k.createCacerts(false); err != nil {
-			log.Infof("Failed to create Cacerts with namespace %s in primary cluster", k.Namespace)
+	var testIstioYaml string
+	// Use the operator manifest when operator mode enabled
+	if *useOperator {
+		istioYaml = authOperatorInstallFile
+		yamlDir := filepath.Join(istioInstallDir, istioYaml)
+		baseIstioYaml := filepath.Join(k.ReleaseDir, yamlDir)
+		testIstioYaml = filepath.Join(k.TmpDir, "yaml", istioYaml)
+		k.fileCopy(baseIstioYaml, testIstioYaml)
+		if err := util.KubeApply("istio-operator", testIstioYaml, k.KubeConfig); err != nil {
+			log.Errorf("Istio operator %s deployment failed", testIstioYaml)
+			return err
 		}
-	}
-	if err := util.KubeApply(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
-		log.Errorf("Istio core %s deployment failed", testIstioYaml)
-		return err
+	} else {
+		// Create istio-system namespace
+		if err := util.CreateNamespace(k.Namespace, k.KubeConfig); err != nil {
+			log.Errorf("Unable to create namespace %s: %s", k.Namespace, err.Error())
+			return err
+		}
+
+		// Apply istio-init and generate an Istio manifest for later
+		yamlDir := filepath.Join(istioInstallDir, initInstallFile)
+		baseIstioYaml := filepath.Join(k.ReleaseDir, yamlDir)
+		testIstioYaml = filepath.Join(k.TmpDir, "yaml", istioYaml)
+		if err := k.generateIstio(baseIstioYaml, testIstioYaml); err != nil {
+			log.Errorf("Generating istio-init.yaml")
+			return err
+		}
+
+		if err := util.KubeApply(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
+			log.Errorf("istio-init.yaml  %s deployment failed", testIstioYaml)
+			return err
+		}
+
+		// TODO(sdake): need a better synchronization
+		time.Sleep(20 * time.Second)
+
+		// Generate main manifest for application in the rest of this function
+		// depending on E2E flags
+		yamlDir = filepath.Join(istioInstallDir, istioYaml)
+		baseIstioYaml = filepath.Join(k.ReleaseDir, yamlDir)
+		testIstioYaml = filepath.Join(k.TmpDir, "yaml", istioYaml)
+
+		if err := k.generateIstio(baseIstioYaml, testIstioYaml); err != nil {
+			log.Errorf("Generating yaml %s failed", testIstioYaml)
+			return err
+		}
+
+		if *multiClusterDir != "" {
+			if err := k.createCacerts(false); err != nil {
+				log.Infof("Failed to create Cacerts with namespace %s in primary cluster", k.Namespace)
+			}
+		}
+		if err := util.KubeApply(k.Namespace, testIstioYaml, k.KubeConfig); err != nil {
+			log.Errorf("Istio core %s deployment failed", testIstioYaml)
+			return err
+		}
 	}
 
 	if *multiClusterDir != "" {
@@ -1162,6 +1234,10 @@ func replacePattern(content []byte, src, dest string) []byte {
 	return content
 }
 
+// This code is in need of a reimplementation and some rethinking. An in-place
+// modification on the raw manifest is the wrong approach. This model is also
+// not portable to our future around IstioControlPlane where we don't necessarily
+// have a manifest to in place modify...
 func (k *KubeInfo) generateIstio(src, dst string) error {
 	content, err := ioutil.ReadFile(src)
 	if err != nil {
